pdf_extraction/requirements.txt
## use python=3.10
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb
faiss-cpu
langchain_google_genai
